chore(gsdk): rename module types to subscription
